// src/Models/IRAIDLogs.ts
export interface IRAIDLogs {
    // id: number;
    SelectType: string;
    RAIDId: string;
    IdentificationDate: string;
    RiskDescription: string;
    AssociatedGoal: string;
    RiskSource: string;
    RiskCategory: string;
    Impact: string;
    RiskPriority: string;
    ImpactValue: Number;
    ProbabilityValue: Number;
    RiskExposure: Number;

}